Remove MUST NOT on 0.00 empty messges
RFC7252 Section 4.1 is clear on that there can not be any data in an
empty message (the name alone should suffice), that includes options --
we would be forbidding something that is impossible, calling people to
ask whether it may be possible for other options.

I not sure but think Carsten pointed this out in some review, but at any
rate this should be an improvement.