CoAP-to-CoAP proxies
I found the CoAP-to-CoAP proxy text a bit strange, it seemed to be limiting CoAP-to-CoAP proxy behaivior very much. I rewrote the paragraph so the CoAP-to-CoAP proxy can act like any server.

The current text still forbids the CoAP-to-CoAP proxy to act as a client when Echo is used. Cannot a CoAP-to-CoAP proxy create its own requests to make sure to keep it's cache fresh? What does such a proxy do when it gets Echo back from the server?