Make normative requirement on not-recycle only per-purpose
The previous requirement read like even clients that support Request-Tag
in principle but do not use it for any particular purpose would be
required to not recycle -- and that requirement could not even be fully
understood as the "except when concluded" clause can not be decided
without the purpose's definition of conclusion.

Rephrased to make the requirement depend explicityly on the use for any
particular purpose.

Addressing a point from Francesca's review