Set details of stateless proxy out of scope
This tones down the stateless-blockwise promises even further, based on
the statement on [1] that we don't need to solve it in full.

An example where it should work is given along with where it's
difficult, and in total it becomes an example of how RT can be used,
without saying anything normative about it, stating it's tricky and out
of scope.

Closes: https://github.com/core-wg/echo-request-tag/pull/59

[1]: https://mailarchive.ietf.org/arch/msg/core/IoIhwW6qUKmq85S2xMOjYUY5i0g