Echo proxy processing: Relax to uphold REST architecture
This allows a proxy to terminate Echo, provided it can uphold the
freshness properties. (Previously, it had to pass the original token
on). This is no severe change as we already allow cross-proxies to
replace one hop's Echo with another freshness mechanism, and does not
harm the use cases. It helps align the option with the general
REST principle of intermediaries being full clients and servers. (Ask
Klaus about the details).